+++ core/server/api_container/server/startosis_engine/kurtosis_instruction/remove_service/remove_service.go
@@ -2,6 +2,8 @@ package remove_service
 
 import (
 	"context"
+	"github.com/kurtosis-tech/kurtosis/api/golang/core/kurtosis_core_rpc_api_bindings"
+	"github.com/kurtosis-tech/kurtosis/api/golang/core/lib/binding_constructors"
 	"github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/service"
 	kurtosis_backend_service "github.com/kurtosis-tech/kurtosis/container-engine-lib/lib/backend_interface/objects/service"
 	"github.com/kurtosis-tech/kurtosis/core/server/api_container/server/service_network"
@@ -53,8 +55,11 @@ func (instruction *RemoveServiceInstruction) GetPositionInOriginalScript() *kurt
 	return instruction.position
 }
 
-func (instruction *RemoveServiceInstruction) GetCanonicalInstruction() string {
-	return shared_helpers.CanonicalizeInstruction(RemoveServiceBuiltinName, kurtosis_instruction.NoArgs, instruction.getKwargs())
+func (instruction *RemoveServiceInstruction) GetCanonicalInstruction() *kurtosis_core_rpc_api_bindings.KurtosisInstruction {
+	args := []*kurtosis_core_rpc_api_bindings.KurtosisInstructionArg{
+		binding_constructors.NewKurtosisInstructionKwarg(shared_helpers.CanonicalizeArgValue(starlark.String(instruction.serviceId)), serviceIdArgName, true),
+	}
+	return binding_constructors.NewKurtosisInstruction(instruction.position.ToAPIType(), RemoveServiceBuiltinName, instruction.String(), args)
 }
 
 func (instruction *RemoveServiceInstruction) Execute(ctx context.Context) (*string, error) {
@@ -67,7 +72,7 @@ func (instruction *RemoveServiceInstruction) Execute(ctx context.Context) (*stri
 }
 
 func (instruction *RemoveServiceInstruction) String() string {
-	return instruction.GetCanonicalInstruction()
+	return shared_helpers.CanonicalizeInstruction(RemoveServiceBuiltinName, kurtosis_instruction.NoArgs, instruction.getKwargs())
 }
 
 func (instruction *RemoveServiceInstruction) ValidateAndUpdateEnvironment(environment *startosis_validator.ValidatorEnvironment) error {